cmd/govim: add testscript tests for new files not on disk
The scenarios are described in more detail in the respective tests, but
these sorts of tests look to specifically assert that we see the correct
behaviour for buffers whose contents have not been saved to disk.

For now we skip scenario_default/quickfix_new_files.txt whilst we await
a fix for golang.org/issues/36661.