cmd/govim: by default don't detect if user is busy in testscript scripts (#694)
In testscript tests we are simulating the user doing things by calls to
the special vim command. Calls to this command are, in effect, a
one-shot thing. That is, we simulate the changes the user to have made
with a single command. Those changes to all intents and purposes take
effect in an instant, i.e. they are not simulating the user pressing
keys over a period of time.

Sometimes these commands can get Vim into a rather confused state with
respect to CursorHold(I)/CursorMoved(I). Indeed we have had to manually
insert some CursorHold events into our tests in order to convince Vim
that we are in fact "done".

Debugging this in Vim will take some time. So the most practical thing
to do in the short term is to disable this "busy" detection by default
in our testscript scripts, and instead have a specific suite of tests
that verify the behaviour under much more controlled circumstances.

We have created #695 as a follow-up for the "busy" suite of tests.