cmd/govim: leaving userbusy always updates text property highlighs (#700)
Diagnostic highlights are always updated when leaving userbusy.
If the buffer changed since the last diagnostic received from gopls,
we might end up with a buffer that doesn't contain the lines/cols
where we try to place text properties. This will thrown an error.

This fix adds the same logic that exists for other diagnostic
features, so that leaving userbusy only triggers text property
placment if there are new diagnostics.