cmd/govim: ensure that GOPATH is in a hidden directory
In testscript tests, Vim is opened with a working directory of the
workdir. This means that the govim file watcher is watching every file
beneath that directory (every file not hidden, that is).

GOPATH is currently $WORK/gopath which means that the govim file watcher
will, incorrectly, be spotting .go files that change within the module
cache.

Fix this be explicitly putting GOPATH within $HOME which is already
defined as $WORK/.home