cmd/govim: skip flakey test for GVim v8.1.1711 (#703)
Per https://github.com/vim/vim/issues/5507#issuecomment-576825067,
scenario_goimports_local/existing_import.txt is known to be flakey for
GVim v8.1.1711. Whilst we understand the underlying reason in that
thread, we choose to skip this test on the understanding the flake is
related to a Vim bug of some sort.