cmd/govim: make all error/warning message checks identical (#712)
Ensure that all of the commented-out error/warning message checks at the
end of testscripts are identical and correct.

Do so by writing a simple helper program, txtarutil, that does this for
us. As part of this make genconfig an internal command along with
txtarutil. Both commands run as part of go:generate directives.

There should be no semantic changes to any tests as part of this commit,
only comments.