cmd/govim: split config.FormatOnSaveGoImports into component parts (#721)
This is a breaking change.

gopls separates the concept of fixing imports (CodeAction) and
formatting (Formatting). We should follow suit on the govim side.

Therefore, FormatOnSaveGoImports changes meaning from being fixing
imports and formatting, to solely fixing imports. Correspondingly,
CommandGoImports also changes its semantics.

We introduce FormatOnSaveGoImportsGoFmt as the combined action and this
then becomes the default for Config.FormatOnSave.