cmd/govim: extract text property removal to own func
This change extracts text property removal to it's own func.

It acts as gardening for implementing references highlights where we'll
start using another text property ID, and therefore it is added as an
internal type.

A new batch method is also added, BatchStartIfNeeded() that is used to
prevent unnecessary round trips to vim when calling a function that
might or might not be within a batch scope already.