tests: add testscript condition for issues
Some of our testscript tests are skipped because they are dependent on
fixes for issues that remain open, i.e. the fixes have not landed. We
try to annotate this skips with an appropriately structured comment,
e.g.:

    skip 'Pending fix for golang.org/issues/36789'

However, this approach makes it hard to not skip the test when you are
looking to verify a fix. We end up having to suggest a sed command (or
similar) to remove/comment out the skip.

Instead we now provide a custom testscript condition that allows issue
tracker links as the condition itself, e.g.

    [golang.org/issues/36789] skip
    [github.com/govim/govim/issues/1234] skip

This means we retain the structure and linkage of the current approach.

Such a condition is always satisfied unless the the
GOVIM_TESTSCRIPT_ISSUES environment variable is a regular expression
that matches the condition in question. For example:

    GOVIM_TESTSCRIPT_ISSUES=.

will cause all issue tracker conditions not to be satsfied, i.e. we
don't skip any tests.