cmd/govim: add test for racey cmd/go updates per golang.org/issues/37164
Test the racing of gopls' usages of go/packages which result in racey usages of
cmd/go modifying go.mod. More detail in:

    https://github.com/golang/go/issues/36772#issuecomment-578933218

and preceding comments. The test below is an exact copy of the setup that triggered
this failure. The go.mod is intentially "dirty" such that any calls to cmd/go
will cause the require directive to be added. One such call happens in the initial
workspace load, the other is triggered by GOVIMGoToDef if the initial call has not
yet completed. Hence the two race to update go.mod