tests: add testsetup.RaceSlowndownFactor (#740)
When we run tests under race mode, all timings are slowed by some
factor. This means that tests which sleep in order to ensure, for
example, that no diagnostics are received, potentially have a higher
chance of passing with false negatives, i.e. they stop looking for
diagnostics too early.

Fix this by introducing a race slowdown factor which is controlled by
the environment variable GOVIM_TEST_RACE_SLOWDOWN. The current
CI-configured default is 1.5

Because some CI systems timeout if no output has been received for a
given period of time, this also requires that we show progress whilst
running the (slower race) builds. We therefore introduce
./internal/cmd/dots which is a simple os/exec wrapper that outputs '.'
progress ticks until all the output from a command is ready, at which
point that output is printed to os.Stdout (the distinction between
stdout and stderr is lost).