cmd/govim: ensure go-to-def tests have stable go.mod files (#736)
golang.org/issues/36772 uncovered the fact the go-to-def tests were
making changes to go.mod files as part of the test. Whilst this
shouldn't be a problem per se, because gopls does not currently handle
cmd/go instances racing on go.mod, it makes sense to avoid this and
instead write a specific test that checks gopls behaves in the correct
way in the presence of multiple cmd/go instanes racing on go.mod.

This commit moves go-to-def tests to have stable go.mod files and
simplifies them in the process by using a module cache definition.

We also switch away from using bufname("") to the more precise
expand('%:p').