cmd/govim: fix embarrassingly bad bug in completion textedit handling (#754)
As part of the current completion code that handles applying of
additional text edits we have:

	for _, c := range v.lastCompleteResults.Items {
		if c.Label == chosen.Abbr {
			match = &c
		}
	}

In a previous change, compltion items changes from being a pointer to a
non-pointer value. Hence we flipped the setting of match from c to &c.
But we forgot the break (which arguably could/should have been there
before). Which means that a further iteration of the for loop after a
match will cause the match to find the text edits from another match.
Specifically, we will be left with the additional text edits from the
last match in the list.

Fix that by breaking out of the loop on a match.

As part of this also fix a test which was incorrectly updated as a
result of this bug (which actually makes the bug itself even worse!)