cmd/govim: add tests for gopls features where files are initially empty (#727)
We have a number of tests that check gopls behaviour where files do not
exist on disk, i.e. they are just buffers in memory. But we are missing
tests to verify that if .go files on disk are empty and then "filled in"
by changes in the editor, that things work as expected.

Add some tests of that sort.