cmd/govim: refactor batch call/expr handling
To better express the fact that sometimes you want to handle when a
batch call fails (rather than asserting it must succeed), we split out
BatchEnd into two methods:

    BatchEnd() ([]json.RawMessage, error)
    MustBatchEnd() []json.RawMessage

MustBatchEnd retains the current semantics of panic-ing if an error is
returned by Vim in the course of evaluating the call/expression.

We extend the assertion capabillity of batch calls/expressions to
support ignoring of certain errors via s:mustBeErrorOrNil. This means
that the assertion functions within the VimScript shim need to be
parameterised. Hence the assertion functions defined in the VimScript
shim, e.g. s:mustBeZero, s:mustBeErrorOrNil and s:mustNoError, are
actually partial functions that return assertion functions.

As part of a related change, the assertion functions themselves now have
the following signature:

    func(v interface{}, err error) (bool, error)

(in Go terms). A non-type(v:none) error value is the caught value in
case an exception was thrown during the evaluation of the
call/expression. The value v is as before, the call return value or
expression result. This therefore allows the caller of a batch to assert
on both the error value and the return value.

Correspondingly, s:mustNothing becomes s:mustNoError, and we modify the
implementation of s:mustBeZero to assert no error value.