cmd/govim: add support for loading an initial file in testscript tests (#785)
PR #784 has uncovered a need to be able to load an initial file as part
of a testscript test. Support this by allowing a vim_config.json file to
be present in a test's archive, a JSON file that is parsed and passed to
the testdriver as Vim-specific config.

Add a test to verify that things work as expected.