cmd/govim: provide env var to limit gopls CPU consumption
In #471 and #755 we have seen instances of govim being sluggish.
In #755 that sluggishness appears to be strongly correlated
with gopls consuming large amounts of CPU during type checking/analysis
phases, especially at startup (where gopls is starting from a cold
cache).

As an experiment we introduce the environment variable
GOVIM_GOVIM_GOMAXPROCS_MINUS_N to limit the amount of CPU gopls can use.
If an integer value n is supplied such that: 0 < n < runtime.NumCPU(),
gopls is run in an environment where GOMAXPROCS=runtime.NumCPU() - n. If
a percentage value p is supplied, e.g. 20%, gopls is run in an
environment where GOMAXPROCS=math.Floor(runtime.NumCPUs() * (1-p)).
govim will panic if supplied with a value of n or p that would result in
a value of GOMAXPROCS such that 0 >= GOMAXPROCS or GOMAXPROCS >
runtime.NumCPU()

Also update govim to log errors returned by tomb.Go functions in order
that errors can be traced (we haven't had to use this before now).

Updates #755