cmd/govim: move unimported completion tests to scenario_default
The default for the unimported completions config option in gopls (and
govim) switched to be "on" some time ago. Hence we do not need a special
scenario for the corresponding tests, and can move them to
scenario_default.

Also make a couple of documentation tweaks to make clear default values.