cmd/govim: track all buffers in govim (#784)
Currently we only track .go files in govim. This makes it impossible,
however, to use commands like GOVIMStringFn in non-go files.

Therefore track all buffers within govim, only relaying events to gopls
for files that matter (currently *.go and go.mod files). We also ensure
that we only try and parse *.go files for govim motion support.

As part of this change we more precisely handle buffer events from Vim.
The corresponding methods on *vimstate in buffer_events.go document
this. Remembering that, per https://github.com/vim/vim/issues/5336, on
Vim startup we do not get buffer events, we therefore manually trigger
govim autocommands to simulate BufNew and BufWinEnter (for loaded
buffers) once govim has finished initialising. We acknowledge, annotate
and workaround a couple of Vim bugs that affect the buffer event
lifecycle. The workarounds can be undone (i.e. strict expectations on
behaviour) by setting GOVIM_VIM_STRICT_BUFFER_LIFECYCLE=true. If these
bugs get fixed we can cleanly undo the workarounds.

A buffer's lifecycle is therefore represented by an entry in
vimstate.buffers, with the Loaded state correctly reflecting whether the
buffer is loaded or not.

As part of this refactor we also tidy up the methods and ordering in
buffer_events.go to reflect the buffer lifecycle.

The refactor of buffer events uncovered a bug in our handling of changed
diagnostics from gopls. Currently we keep track of whether quickfix,
sign and highlights have changed since we last tried to update any of
the aforementioned categories. However, this is imprecise: we should
instead keep track of the diagnostics that were last used to update Vim.
Therefore we now track the diagnostics used to, for example, update
diagnostics via govim.lastDiagnosticsQuickfix, of type
*[]types.Diagnostic.

The refactor also necessitated a change in how we determine whether two
quickfix entries are equal or not. We now compare quickfix entries
modulo the buffer number (which can, and does, change in a Vim session).

We also add some context to each fire of an autocommand, such that the
call from Vim -> govim that is made in response to an event includes the
autocommand event definition so that it's easier to spot what event is
firing. e.g:

   recvJSONMsg: ["function","autocommand:3","BufNewFile,BufRead *"...

As part of ensuring that GOVIMStingFn does indeed work in non .go files,
we add a test to stringfn.txt to confirm as much.

Because we now more reflect buffer loaded state with open/close events
to gopls, we now see a different number of diagnostic events. We update
one test that has such assertions.

We also add a number of tests to verify that govim behaves when saving
unnamed buffers to go/non-go files, opening go/non-go files and
switching to go/non-go files.