cmd/govim: handle buffer name changing in BufWinEnter (#798)
A buffer name can change as a result of BufReadCmd. Some plugins use
BufReadCmd to handle special reads of files. For example,
github.com/plan9-for-vimspaceplan9-for-vimspace adds an address handler
to BufReadCmd to handle an attempt to open /path/to/file.go:5:6, i.e. at
a specific address (line:col). Most such handlers then fire a BufRead and
BufWinEnter to complete the buffer life-cycle events, at which point we
can correct the name of the buffer.

There is currently no test for this, but it has been verified locally.