all: back-out changes to track all buffers
This commit was formed via the following sequence of commands:

    git diff -R b989e956 | git apply
    git show ec0360d5 | git apply
    git show 9f146506 | git apply
    git show c7ecca29 | git apply
    git show 6b6b570a | git apply
    git show a3741ae9 | git apply

Any manual tweaks required to re-add small changes we need from 8826ef5f
and friends will be added in later subsequent commits.

The more important diff to look at here is:

https://github.com/govim/govim/compare/b989e956...backout_all_buffer_tracking

i.e. that with respect to b989e956 (the first commit prior to the "track
all buffers" changes) the only diffs are changes that are _not_ related
to "track all buffers"