cmd/govim: handle nil response to Hover
In the current gopls implementation, if there is nothing under the
cursor when we call Hover we simply get a zero value protocol.Hover
returned (albeit via a pointer variable).

As of de023d59a5d1 gopls now returns a zero value of type
*protocol.Hover. Hence where we previously assumed we would not get a
nil value we now do.

Fix that.