cmd/govim: fix handling of changed diagnostics from gopls (#811)
In 8826ef5f (since backed out) we uncovered a bug in our handling of
changed diagnostics from gopls. Currently we keep track of whether
quickfix, sign and highlights have changed since we last tried to update
any of the aforementioned categories. However, this is imprecise: we
should instead keep track of the diagnostics that were last used to
update Vim. Therefore we now track the diagnostics used to, for example,
update diagnostics via govim.lastDiagnosticsQuickfix, of type
*[]types.Diagnostic.

The refactor also necessitated a change in how we determine whether two
quickfix entries are equal or not. We now compare quickfix entries
modulo the buffer number (which can, and does, change in a Vim session).

We also fix a bug that was also hidden by the current implementation
whereby signs weren't being re-placed when a buffer is reloaded.

We also add some context to each fire of an autocommand, such that the
call from Vim -> govim that is made in response to an event includes the
autocommand event definition so that it's easier to spot what event is
firing. e.g:

   recvJSONMsg: ["function","autocommand:3","BufNewFile,BufRead *"...

We also fix up GOVIMStringFn to work correctly at the end of files and
update the test accordingly to add coverage of this.

We also add a number of tests to verify that govim behaves when saving
unnamed buffers to go/non-go files, opening go/non-go files and
switching to go/non-go files, skipping the one case that does not work
in light of us having raised #810.