cmd/govim: remove one of the file watchers (#819)
There are currently two file watchers being started, one in Init(...)
and one in startGopls(...). This leads to duplicate
DidChangeWatchedFiles calls to gopls.

Remove one of them, as well as a dated comment about this being a
temporary fix.

This change also removes a field "files" (on the modwatcher struct)
that isn't used anymore.