cmd/govim: safely handle busy state changes for non-.go buffers
In 8926deca we backed out the changes for tracking all buffer changes.
But in so doing we didn't properly handle the situation where the busy
state changes and the cursor is in a non-go file.

Fix that and add a test to verify we don't regress.

Whilst we are at it:

* consistently use GOVIM_test_SetUserBusy in our testscript tests
* slightly refactor the types/consts used for getting the current cursor
  position

Fixes #821