cmd/govim: add experimental autoread loaded buffers support (#826)
If a file is updated on disk while it is open in vim, govim/gopls won't
pick up the change automatically.

This change adds a config setting "ExperimentalAutoreadLoadedBuffers"
that is used to reload buffers for files that change outside vim. It
requires the buffer to be "loaded" since it uses "checktime" to reload
the file.

This feature is not enabled by default. Add the following to your vimrc
to enable it:

set autoread
set hidden
call govim#config#Set("ExperimentalAutoreadLoadedBuffers", 1)

"set autoread" will suppress a confirmation prompt when a buffer that
isn't modified is reloaded (a prompt will still be shown if the buffer
is modified). This line is optional.

"set hidden" will configure vim to keep buffers loaded when switched,
instead of the default that is to abandon them. This is required for the
feature to work with buffers that aren't switched away.