cmd/govim: persist highlights for explicit GOVIMHighlightReferences call
When manually invoking GOVIMHighlightReferences, it makes sense to
persist the highlights for as long as we can. To be safe we assume that
making any change to any .go file will potentially invalidate the
highlighted references (we might relax this at a later stage). We also
add the GOVIMClearReferencesHighlights command to explicitly remove the
highlights.

As part of this same commit we:

* move the definition of currentReferences to vimstate where it belongs
  (it is only ever accessed on the vimstate "thread")
* Tidy up and simplify some logic regarding removal of highlights