cmd/govim: add a first cut of signature help (#854)
GOVIMSignatureHelp shows a popup with signature information and
documentation for the function or method call enclosed by the cursor
position. The cursor must be after the left parentheses of the call
expression. If there is no signature help available for the current
cursor position, no popup is shown. Moving the cursor or the mouse
causes the popup to be dismissed.