cmd/govim: fix completion candidate additional text edits for dups
In e10617c6 we permitted "duplicate" candidates in completion results.
(These items are not in fact duplicates because they are distinguished
by their detail.) However, we failed to fix the logic that matches the
chosen completion candidate versus the list of options; we also need to
include the detail.

Fix that and track adding a test in #867.