cmd/govim: fix scheduling of work (#936)
Currently, when work is scheduled by govim in Vim, it's possible for the
work to execute "inline" immediately if it is safe to do so, before the
schedule call returns. If a second schedule call happens just after the
first (but before the immediate execution of the first bit of work has
started), the second piece of work will be added to the scheduled work
queue. When the first piece of work comes to be executed, there are now
two bits of work in the queue. Our "draining" of the scheduled work
queue works through them both. However, at this point the first schedule
call is still blocked. So if the second piece of work blocks pending the
completion of the first piece of work then we hit a deadlock, because
the schedule of the first bit of work is blocked on the compeltion of
the second piece of work.

This commit fixes that by never "inlining" the immediate execution of a
scheduled piece of work: it is always run asynchronously from the
schedule call itself.

As part of this, fix a bug in the govim tests where our echom calls were
blocking in Vim world with a user prompt (unclear why we weren't running
into this before).