cmd/govim: remove internal buffer on BufWipeout
There are cases where we might get a BufWipeout autocmd without the
BufDelete (most notably when running :vimgrep). When that happen we kept
the internal buffer in govim even if we had no use of it anymore.

When using :bwipe we do get both BufDelete and BufWipeout, but as far as
we know BufDelete will always appear before BufWipeout. Trying to remove
an internal buffer due to BufWipeout will not cause an error.

Fixes #1033