cmd/govim: check quickfix list origin before update (#1058)
QuickfixDiagnostics routine updates the quickfix list each time it
receives new diagnostics from gopls. This can be problematic if the
quickfix list has been populated by an other plugin, because the user
could be using this plugin.

The fix uses the quickfix title to determine if the quickfix is "owned"
by govim or by an other plugin. If the quicfix title is empty (which
means the quickfix has never been used before) or equal to "govim", or
if the quickfix itself is empty, then govim considers it can change it,
else it considers it cannot.

This condition is ignored when :GOVIMQuickfixDiagnostics is invoked by
the user, making possible for the user to display again the diagnostics
when he wants to.