cmd/govim: fix shutdown sequence (#1062)
Currently we do not properly initiate shutdown from Vim when a user
wants to exit. Instead, we simply close the channel to govim. This means
that Vim does not (indeed cannot) wait for govim to complete its
shutdown. Instead, the govim process (and the child gopls process)
simply get terminated.

Fix this by initiating a proper shutdown sequence when Vim is about to
exit. This causes Vim to block until the govim plugin Shutdown call is
complete. This allows govim and gopls to properly tidy up.

This also means that it is the responsibility of Vim to initiate
shutdown, hence we can remove som redundant logic in govim.go where the
plugin tried to cleanup after itself.

Fixes #842