cmd/govim: add opt-in logging of raw file watcher event
The file watcher can be tricky to debug due to the unstable nature of
file watchers. Logging raw events in CI could have been very helpful
when we tried to find previous flakes and other bugs.

This change adds the possibility to add raw file watcher event logging
via GOVIM_LOG, e.g.:

$ export GOVIM_LOG=on,watcher

We keep it as an opt-in since it could generate a lot of log entries for
certain user workflows, but do enable it in CI since we control the
tests.