cmd/govim: do not wait more than 500ms for gopls Shutdown
As explained in golang.org/issue/45476, all call to Shutdown does not
properly interrupt "work" that might be in progress in gopls. This has
the effect of blocking an exit from Vim.

Workaround this for now by limiting the wait time to 500ms, which is
effectively the measure of a user's patience before anger/frustration
sets in!