testdriver: simplify condition version logic
vim/gvim conditions can currently take these forms:

    vim          # only run if vim
    gvim         # only run if gvim
    vim:v1.2.3   # only run if vim >= v1.2.3
    gvim:v1.2.3  # only run if gvim >= v1.2.3

The version specifier prefix is superfluous however, because we could
simply use v1.2.3 to mean the version of the editor, and then the plain
vim/gvim to specify the flavour.

Doing so helps cut down the number of exceptions we need to write in
case a later version of Vim (or Gvim) introduces a breaking change
(typically in the marshalled encoding of a function return value).

Refactor condition logic and add some tests.