plugin/govim: drop .log suffix on vim channel log (#1103)
Some flavors of "mktemp", at least macOS and possibly OpenBSD, require
that the template ends with X's (instead of just having them anywhere in
the template). On macOS they won't be converted to a random number at
all.

This change drops the .log extension from the vim channel log. We could
have run mktemp in "dry-run" mode to just generate a file name and
append the extension afterwards, but that creates and removes the file
on macOS. Doing that introduces a race that makes it fairly easy for
another users on the same system to "hijack" the logfile.