plugin: fix "git rev-parse" when running under "git commit" (#1115)
I had been seeing an issue where running "git commit" with vim as the
editor would almost always get slowed down with:

	Installing govim and gopls

It did not make a lot of sense to me; my version of Go and govim was not
changing between the commits. I was simply editing an unrelated Go
project and committing the changes in between.

Even more weirdly, it seemed like this issue only occurred when I ran
"git commit" inside a checked out git submodule.
It did not happen in a direct git clone.

I added GOVIM_DEBUG_INSTALL to allow myself to debug what was going on.
All the variables looked right, except "commit"; it was pointing to the
commit I was in the process of committing, not govim's HEAD.

It turns out "git commit" was setting environment variables like GIT_DIR,
which were making "git rev-parse HEAD" not use plugindir as the repo.

Call "git rev-parse" with "env -i" to run it with an empty environment.
This should prevent future problems with GIT_ env vars set when vim runs.
It seems safe, as "git rev-parse" is a fairly simple operation.

I leave GOVIM_DEBUG_INSTALL in place, to make future debugging easier.