cmd/govim: support setting gopls memory mode
In gopls there is an experimental feature that can be used to reduce the
memory footprint by degrade features in packages without any open files.
This can be useful in large single-moduled mono repos that otherwise
consume lots of resources.

We support the both modes defined in gopls, "Normal" and "DegradeClosed"
where "Normal" is the default behaviour. To use "DegradeClosed" add a
line to your .vimrc:

call govim#config#Set("ExperimentalGoplsMemoryMode","DegradeClosed")