all: add support for go1.18beta1
The convention used in txtar archive filenames is to call pre go1.18
expectations blah.pre_go1.18.golden, and the go1.18 expectations
blah.golden. This way, when go1.18 becomes the oldest version in the
build matrix we can simply delete files which have pre_go1.18 in the
name, as well as commands that reference them, with no further renaming
required.

Also:

* make latest version of Vim in build matrix v8.2.4213
* bump to go1.16.13 and go1.17.6
* add convenience environment variable GOVIM_UPDATE for updating
  testscript expectations