deps: update x/tools and gopls to 53efd30f4782
Including changes to adapt to internal API changes in vendored
code, as well as fixing up tests.

Also switch to using an installed version of Vim rather than relying on
Docker-based tests. Update the script to accept the default version,
which is ignored when run via cron in order to test the latest version
of Vim.

Upgrade Go and Vim versions to current latest.